Restore documentation version footer to v0.9.5
