+++ docs/adr/ADR-0012-coreops-package.md
@@ -1,4 +1,4 @@
-# ADR-0012 — Adopt c1c_coreops package; deprecate shared/coreops_*
+# ADR-0012 — Adopt c1c_coreops package; deprecate legacy shared CoreOps shims
 
 ## Status
 
@@ -10,7 +10,7 @@ Accepted
 
 ## Context
 
-CoreOps helpers lived in `shared/coreops_*` alongside infrastructure modules.
+CoreOps helpers lived in legacy shared CoreOps modules alongside infrastructure code.
 That layout made it difficult to share the cog, RBAC checks, and embed renderers
 across bots without dragging the entire `shared` package into new runtimes.
 Keeping the code dual-homed between `modules/coreops/` scaffolding and shared
@@ -19,8 +19,8 @@ files also risked drift once we begin migrating imports for other bots.
 ## Decision
 
 Create an internal package at `packages/c1c-coreops/src/c1c_coreops/` and copy
-the existing CoreOps modules into it unchanged. Leave the `shared/coreops_*`
-files as deprecated re-export shims for one release so current importers keep
+the existing CoreOps modules into it unchanged. Leave the legacy shared
+shims as deprecated re-exports for one release so current importers keep
 working while we stage the follow-up import rewrite.
 
 ## Consequences
@@ -32,4 +32,4 @@ working while we stage the follow-up import rewrite.
 * We must follow up with a PR to update imports to `c1c_coreops.*` and remove
   the shims once callers move over.
 
-Doc last updated: 2025-10-23 (v0.9.5)
+Doc last updated: 2025-10-26 (v0.9.5)
