+++ packages/c1c-coreops/src/c1c_coreops/cache_public.py
@@ -1,0 +1,21 @@
+"""Public CoreOps cache telemetry facade."""
+
+from __future__ import annotations
+
+from shared.cache.telemetry import (
+    CacheSnapshot,
+    RefreshResult,
+    get_all_snapshots,
+    get_snapshot,
+    list_buckets,
+    refresh_now,
+)
+
+__all__ = [
+    "CacheSnapshot",
+    "RefreshResult",
+    "get_all_snapshots",
+    "get_snapshot",
+    "list_buckets",
+    "refresh_now",
+]
