+++ packages/c1c-coreops/tests/test_coreops_import_no_env.py
@@ -6,7 +6,7 @@ import sys
 
 
 def _ensure_src_on_path() -> None:
-    root = pathlib.Path(__file__).resolve().parent.parent
+    root = pathlib.Path(__file__).resolve().parents[3]
     root_str = str(root)
     if root_str not in sys.path:
         sys.path.insert(0, root_str)
